backend/requirements.txt
fastapi[standard]
pymongo
python-dotenv
pwdlib[argon2]
python-jose[email]
python-multipart
pydantic-settings
email-validator
pymupdf
openai
reportlab